feat: add soft delete to UserDatabaseService and modify IpServicesImpl to  search for an IP address across multiple subnets
feat: Add a soft delete function to UserDatabaseService. This prevents deleting a user from the database and simply marks them as deactivated.

refactor: Modify IpServicesImpl to search for an IP address across multiple subnets, allowing for easy addition of multiple subnets in the future.

refactor: Modify UserEntity, PeerEntity, and IpEntity so that the table name in the database is defined using the annotation @Table(name = ""). This results in a cleaner implementation when creating @Query.